requirements.txt
streamlit==1.29.0
requests==2.31.0
python-jenkins==1.8.0
PyGithub==1.59.1
ollama==0.1.7
python-dotenv==1.0.0
pytest==7.4.3
subprocess32==3.5.4
gitpython==3.1.40 